requirements.txt
asgiref
Django
django-ranged-response
django-simple-captcha
django-simpleui
lxml
mysql-connector-python
mysqlclient
pillow
pip
PyMySQL
python-docx
setuptools
sqlparse
typing_extensions
tzdata
wheel
jieba

faker
pandas
numpy
wordcloud